// src/components/buttons/buttons.interface.tsx
export type ButtonColorType =
  | "inherit"
  | "primary"
  | "secondary"
  | "success"
  | "error"
  | "warning"
  | "info";

export type ButtonVariantType = "text" | "outlined" | "contained";

export type ButtonSizeType = "small" | "medium" | "large";
// ...
export interface LinkButtonPropsI {
  color?: ButtonColorType;
  variant?: ButtonVariantType;
  link?: string;
  customStyles?: any;
  children?: ReactNode;
  size?: ButtonSizeType;
  primary?: boolean;
  startIcon?: ReactNode | undefined;
  endIcon?: ReactNode | undefined;
  linkProps?: LinkProps;
  fullWidth?: boolean;
}